+++ RavenMigrations/NullLogger.cs
@@ -1,0 +1,17 @@
+﻿namespace RavenMigrations
+{
+    public class NullLogger : ILogger
+    {
+        public void WriteInformation(string format, params object[] args)
+        {
+        }
+
+        public void WriteError(string format, params object[] args)
+        {
+        }
+
+        public void WriteWarning(string format, params object[] args)
+        {
+        }
+    }
+}